commit pre handling abort and batch commit

+++ sdk-core/src/main/java/dk/ku/di/dms/vms/sdk/core/operational/VmsTransactionTaskResult.java
@@ -1,6 +1,9 @@
 package dk.ku.di.dms.vms.sdk.core.operational;
 
+/**
+ * Placeholder so the scheduler can identify whether a task has failed
+ */
 public record VmsTransactionTaskResult(
-    int tid,
+    long tid,
     int identifier,
     boolean failed){}
